refactor: move st components to core
