+++ core/__init__.py
@@ -1,18 +1,20 @@
-import _1_ytdlp
-import _2_asr
-import _3_1_split_nlp
-import _3_2_split_meaning
-import _4_1_summarize
-import _4_2_translate
-import _5_split_sub
-import _6_gen_sub
-import _7_sub_into_vid
-import _8_1_audio_task
-import _8_2_dub_chunks
-import _9_refer_audio
-import _10_gen_audio
-import _11_merge_audio
-import _12_dub_to_vid
+from . import (
+    _1_ytdlp,
+    _2_asr,
+    _3_1_split_nlp,
+    _3_2_split_meaning,
+    _4_1_summarize,
+    _4_2_translate,
+    _5_split_sub,
+    _6_gen_sub,
+    _7_sub_into_vid,
+    _8_1_audio_task,
+    _8_2_dub_chunks,
+    _9_refer_audio,
+    _10_gen_audio,
+    _11_merge_audio,
+    _12_dub_to_vid
+)
 from .utils import *
 from .onekeycleanup import cleanup
 from .delete_retry_dubbing import delete_dubbing_files
@@ -23,9 +25,6 @@ __all__ = [
     'update_key',
     'cleanup',
     'delete_dubbing_files',
-    'translate_chunk',
-    'translate_lines',
-    'translate_lines',
     '_1_ytdlp',
     '_2_asr',
     '_3_1_split_nlp',
